pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "dishka>=1.7.2",
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "mypy>=1.16.1",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.12.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
    "openai",
    "pydantic[email]>=2.11.7",
    "pyjwt[crypto]>=2.10.1",
]
